requirements.txt
Flask==3.0.2
quote==3.0
google-genai==1.33.0